+++ materials/models/models_kit/xmas/xmastree_miscA_skin2.vmt
@@ -1,0 +1,6 @@
+"vertexlitGeneric"
+{
+	"$basetexture" "models/models_kit/xmas/xmastree_miscA_skin2"
+	"$alphatest" "1"
+	"$nocull" "1"
+}
